Terminate 0.6-dev
Apparently some kind of error happened during the porting of the whole
library, so it doesn't behave correctly. I personally don't have the
knowledge nor the experience to understand where this parser is failing.
The only thing I verified is that the issue is not related to the
modules, but to how the various components of the parser interact with
each other.